+++ package.json
@@ -47,6 +47,7 @@
 		"compression": "^1.7.2",
 		"cookie-universal-nuxt": "^2.0.14",
 		"cors": "^2.8.5",
+		"cron": "^1.8.2",
 		"dotenv": "^6.2.0",
 		"dumper.js": "^1.3.1",
 		"express": "^4.17.1",
